FileStorage/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
python-environ
django-cors-headers
django-rest-framework
django-widget-tweaks
djangorestframework
djangorestframework-jwt
